Implement Distance Calculation and Cluster Point Distribution
The code is now able to find the distance from the center point of each cluster and also able to distribute itself in different clusters depending upon the distance. Also we were able to find the new center points. Making our cluster points more accurate. Some more work is needed to perform this operation until the optimal solution is not found.